One script to rule them all
- Run img/mov/sketch run during the day
- Switch to 'breathing' over night

+++ vrn-run.py
@@ -1,72 +1,30 @@
-#! /usr/bin/env python
-
-import sys, os
-from datetime import datetime
+import os, random
 
 # hide the cursor
 os.system('DISPLAY=:0 xdotool mousemove 800 800')
 
-def main():
-    if shutdown_at_nighttime():
-        sys.exit()
-
-    import random
-
-    # We run via cron every 24 minutes, so this means we only change infrequently
-    if random.randint(0,19) > 17:
-        print("No change")
-        sys.exit()
-
-    import os
-
-    # TODO: select scripts based on this? or dim screen...
-    day_part = get_part_of_day(datetime.now().hour)
-
-    sketches = [ 
-        'colour_wipes',
-        'purple_squares',
-        'special_branch_spotlights',
-        'floyd_bigger',
-        'special_branch_layered_squares',
-        'shapes_grid',
-        'easel_strokes',
-        'led',
-        'spots',
-        'plain_waves',
-        'special_branch_shoreline',
-        'coloured_squares',
-        'mcferrin_circles',
-        'special_branch_natural_squares',
-    ]
-
-    sketch = 'DISPLAY=:0 /usr/local/bin/processing-java --sketch="/home/pi/src/vrniture/processing/{0}/" --run'.format(random.choice(sketches))
-
-    now = datetime.now()
-    print(now.strftime("%Y-%m-%d %H:%M:%S"))
-    print(sketch)
-    print('-----')
+print("vrn-run.py")
+print("----------")
 
 
-    # Stop any existing sketches or films
-    os.system('killall omxplayer')
-    os.system('killall java')
+media = [
+    'img-run.py',
+    'mov-run.py',
+    'sketch-run.py',
+]
 
-    os.system(sketch)    
+# chance cubes!
+if (random.choice([1,2,3,4,5,6]) <= 3):
+    print( '/usr/bin/env python /home/pi/src/vrniture/' + random.choice(media) + ' >> /tmp/cron.log 2>&1' )
+    os.system( '/usr/bin/env python /home/pi/src/vrniture/' + random.choice(media) + ' >> /tmp/cron.log 2>&1' )
 
-def get_part_of_day(hour):
-    return (
-        "morning" if 5 <= hour <= 11
-    else
-        "afternoon" if 12 <= hour <= 17
-    else
-        "evening" if 18 <= hour <= 22
-    else
-        "night"
-    )
+    import datetime
+    if (datetime.datetime.now().hour >= 22):
+        # stop any existing sketches or films
+        os.system('killall omxplayer')
+        os.system('killall java')
+        os.system('killall feh')
 
-def shutdown_at_nighttime():
-    if get_part_of_day(datetime.now().hour) == 'night':
-        return True
-
-if __name__ == "__main__":
-    main()
+        os.system('DISPLAY=:0 /usr/local/bin/processing-java --sketch="/home/pi/src/vrniture/processing/breathing/" --run >> /tmp/cron.log 2>&1')
+else:
+    print 'Pass'
